+++ app/src/main/java/com/maurodev/chronometer/ui/principal/PrincipalViewModel.kt
@@ -1,0 +1,4 @@
+package com.maurodev.chronometer.ui.principal
+
+class PrincipalViewModel {
+}